# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Get all products",
        "responses": {
          "200": {
            "description": "Successfully retrieved all products"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Create a new product",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new product"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Edit an existing product",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/editProduct"
            }
          },
          {
            "name": "productId",
            "in": "path",
            "description": "ID number of a specific product",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully edited the product"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/products/{productId}": {
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Soft delete an existing product",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID number of a specific product"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the product"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/admin/products": {
      "get": {
        "tags": [
          "Admin-Products"
        ],
        "description": "Get all products",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all products"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin-Products"
        ],
        "description": "Create a new product",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new product"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createProduct": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Product Example"
        },
        "Description": {
          "type": "string",
          "example": "Product description ...."
        },
        "UnitPrice": {
          "type": "number",
          "example": 15
        },
        "ImageURL": {
          "type": "string",
          "example": "http://images.restapi.co.za/products/product-apple-tv.png"
        },
        "Quantity": {
          "type": "number",
          "example": 5
        },
        "BrandName": {
          "type": "string",
          "example": "Example Brand"
        },
        "CategoryName": {
          "type": "string",
          "example": "Example Category"
        }
      },
      "required": [
        "Name",
        "Description",
        "UnitPrice",
        "ImageURL",
        "Quantity",
        "BrandName",
        "CategoryName"
      ]
    },
    "editProduct": {
      "type": "object",
      "properties": {
        "NewName": {
          "type": "string",
          "example": "Edited Product"
        },
        "NewDescription": {
          "type": "string",
          "example": "Edited Product description ...."
        },
        "NewUnitPrice": {
          "type": "number",
          "example": 15
        },
        "NewImageURL": {
          "type": "string",
          "example": "http://images.restapi.co.za/products/product-apple-tv.png"
        },
        "NewQuantity": {
          "type": "number",
          "example": 5
        },
        "NewBrandName": {
          "type": "string",
          "example": "Edited Brand"
        },
        "NewCategoryName": {
          "type": "string",
          "example": "Edited Category"
        }
      },
      "required": [
        "NewName",
        "NewDescription",
        "NewUnitPrice",
        "NewImageURL",
        "NewQuantity",
        "NewBrandName",
        "NewCategoryName"
      ]
    },
...
  }
}